JVMCBC-167: Allow to disable buffer pooling.
Motivation
----------
It might pay off in the future to have a "kill" switch to disable
buffer pooling. This can be used as a temporary workaround if a
leak in the client is suspected (and not in the user level code).

Modifications
-------------
Add the environment option which is enabled by default. The endpoint
selects the proper allocator based on the environment setting.

Result
------
As a last resort it is possible to disable buffer pooling, but of
course not recommended.

Change-Id: I9517e8e0cbc93e6480efceb058f867125d9752ac
Reviewed-on: http://review.couchbase.org/48104
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>