JVMCBC-168: Deactivate Nagle's Algorithm
Motivation
----------
Nagle's Algorithm in TCP consist in buffering small packets and sending
them as a single packet when the total size vs TCP overhead ratio gets
small.
In terms of latency, in current networks, it is less and less relevant
and introduces latency.

Modifications
-------------
AbstractEndpoint now builds the socket with TCP_NODELAY set to true,
which disables Nagle's algorithm in Netty.

Result
------
Lower latencies are to be expected thanks to this tuning. Note that the
algorithm is now by default disabled in Netty.

Change-Id: Ica335a11bd170bea211e522aa2d0392e8c980581
Reviewed-on: http://review.couchbase.org/48112
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>