JVMCBC-170: Apply a hard lower limit for pool sizes.
Motivation
----------
For various reasons (only one core, wrong user config) it can happen
that the io and/or computation pool sizes get set to a very low
value which can cause deadlocks or resource starvation.

Modifications
-------------
This changeset applies a lower limit of 3 which is a resonable
default lower limit. It also prints INFO level messages to help
the user identify any issues.

Result
------
More resiliency out of the box for different user setups and
environments.

Change-Id: Ic2db596c0ece860fbd8adf263d813b71a3a94fb1
Reviewed-on: http://review.couchbase.org/48171
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>