JVMCBC-171: Make sure nodesExt without hostname is backwards compatible.
Motivation
----------
Couchbase Server versions prior to 3.0.2 do not include the hostname
in the nodesExt config part. In 2.1.1 a regression was introduced
which expected it to be there and therefore broke backwards
compatibility.

Modifications
-------------
The code now grabs the hostname from the nodes section if it is
not available in nodesExt. A test has been added to make sure it
works properly.

Result
------
Backwards compatbible behaviour with server versions older than
3.0.2.

Change-Id: I89b3230826103fb6dc1da94d72674ff23366d22e
Reviewed-on: http://review.couchbase.org/48238
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>