JVMCBC-173: Fix DCP compatibility with transient endpoints.
Motivation
----------
Somewhere between 1.1.0 DP and GA some code was reworked and parts
of the experimental DCP support was broken.

This changeset fixes the compatibility.

Modifications
-------------
To make it work again, a few things had to be put into place:

	- Add a LazyService to connect lazily but limit to 1 connection
	  and then reuse it.
	- Remove the transient nature from the dcp endpoint
	- Add some more intelligent cluster version detection to also
	  now run the DCP int tests when possible.
	- Fix the integration test with an Rx test subscriber

Result
------
Basic DCP integration is now working again, but more changes are
anticipated to make the behaviour more correct and usable.

Change-Id: I81596837033c15db0cac90a869db814f03273b30
Reviewed-on: http://review.couchbase.org/48680
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>