JVMCBC-175: Always Observe master to detect concurrent modifications.
Motivation
----------
When the CAS value changes on the master to the one that should be
observed, there is a chance that the one the code is looking for
will never be replicated (since the server can deduplicate items
and therefore the op will never end up in this form on the
replica).

The code already has a check for that, but it is only executed
if a response from the master is returned. If the user only
specifies ReplicateTo, the master is never asked and therefore
the operation will time out.

Modifications
-------------
To avoid this case, the master is now always included in the poll
operation, and the response is used for a cas sanity check. Of
course if not needed it does not participate in the requirements
checks until success is returned, but it can throw an exception
if the concurrent modification is detected.

The message has also been improved to be more user friendly.

Result
------
Correct behavior on cas concurrent modification detection even when
only replication constraints are needed.

Change-Id: I70ce492b74e23f879a50172f602ff4a4b787972d
Reviewed-on: http://review.couchbase.org/48845
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>