JVMCBC-177: Proactively ask for a new config on failed connect attempts.
Motivation
----------
Currently, the client only proactively grabs a new config once a channel
is set to inactive. If the server respons with a still outdated config,
the client needs to grab a new config at a later point in time.

Modifications
-------------
The SDK now also tries to grab a configuration after every failed connect
attempt, which will eventually be every 4 seconds by default (which is the
upper retry limit). This provides an additional mechanism for the SDK to
get back into a stable state with an up-to-date config.

Result
------
The SDK is now able to load a new config even if a good one is available
after a longer timeframe, for example if a node is shut down and the failover
is initiated much later.

Change-Id: I14201852e8586b823a8051f6f278b18319893e0f
Reviewed-on: http://review.couchbase.org/49102
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>