JVMCBC-180: Avoid #merge when only 1 observable needed in Observe.
Motivation
----------
When it is determined that only one observe call needs to be sent,
the Observable#merge() is not needed, since it is enough to just
pass the one observable downstream.

Modifications
-------------
Conditional block added to check for the list size and if one,
just return the observable.

Result
------
Not much, but helps a little since merge is quite expensive.

Change-Id: Ie428fe0586c3467bd3f492dae356e8d64e906b22
Reviewed-on: http://review.couchbase.org/49596
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>