JVMCBC-181: Refactoring of KeyValueLocator
Motivation
----------
The locator is on the hot code path and the JIT logs report that
especially the locateForCouchbaseBucket method is too hot to be
inlined.

Modifications
-------------
The code has not been altered, but just split up into smaller functions,
making it possible for the JIT to inline it and therefore
possibly make it faster.

Result
------
The code is now more cleanly separated and also it makes it possible
for the JIT to optimize it.

Change-Id: I79256122ea20e45dbdb415161f899f15e2cd4631
Reviewed-on: http://review.couchbase.org/49597
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>