JVMCBC-183: store server's error code in response
Motivation
----------
If and when the server sends a response that has an unknown code, SDK
translates it to ResponseStatus.FAILURE but loses the original status
code.

Modifications
-------------
Added a field to BinaryResponse for storing the original status code.
Initialized via constructor, so all child classes have it added to the
respective constructors, and unit tests reflect that.
Added a note on ResponseStatus.FAILURE's javadoc to indicate that it
denotes an "unknown" status code, and that said code can be found in
the BinaryResponses.

Result
------
Better possibilities for the clients to exactly see what happened in
case of error on the server side when the error code is not yet mapped
in the SDK.

Change-Id: Iddd05fbbfeef1bebcc8288f87b4425ef71710e60
Reviewed-on: http://review.couchbase.org/49732
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>