JVMCBC-185: Make sure password is passed on to the next config.
Motivation
----------
When a new configuration arrives (for example through http streaming
on a memcached bucket) it is important that the bucket password
is passed along from the old to the new config so it does not
get lost.

Modifications
-------------
The code now makes sure to always - before swapping it out - transfer
the password over to the new config.

Result
------
Avoiding authentication issues on node adds through the memcached
http streaming code path.

Change-Id: I0abb3fd0b3a583f554d9adf08e972c2cfc644b71
Reviewed-on: http://review.couchbase.org/49857
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>