JVMCBC-186: Graceful skipping on failing bootstrap node list elems.
Motivation
----------
When only a subset of the nodes in the bootstrap list is failing (for
example because the IP address has an invalid format or the hostname
can't be looked up), the code should not give up but rather ignore
the error and keep moving with the nodes it has.

Modifications
-------------
Errors are now logged and only if the list is empty at the end a
ConfigurationException is thrown.

Result
------
The code is now more resilient to partial invalid hosts in the
bootstrap list and keeps going in a best-effort manner while still
reporting all issues and failing eventually if it can't make
progress.

Change-Id: I7e72a88763ded71a407257055bf4f92b4f1c3c7c
Reviewed-on: http://review.couchbase.org/49900
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>