JVMCBC-184: Prevent race condition on topology change with >1 buckets.
[Manual Backport]

Motivation
----------
When more than one bucket is open at the same time a race condition combined
with our "safety rescheduling" logic prevents the system from getting back
into a stable state.

The two configurations for the buckets race each other with the reconfiguration
logic and the operations always get rescheduled.

In addition, the retry logic needs to be in place for memcached buckets as well,
since the gap can be there present too.

Modifications
-------------
The change is quite trivial in that the rescheduling logic is now only executed
as a fallback when the regular node could not be selected beforehand. This allows
the operations to pass while just to retry the subset which is affected by
the temporary topology difference between the config nodes and the actual
instantiated node list.

Also the retry logic is applied for memcached buckets.

Result
------
More stable behaviour during rebalance when more than one bucket is open.

Change-Id: I80d597bb737bad190fa7a054167da90006e0d1c6
Reviewed-on: http://review.couchbase.org/49938
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>