JVMCBC-191: Log unknown status codes.
Motivation
----------
When the server returns an unknown status code, a log message with the
actual code needs to be written in order to properly debug the situation.

Modifications
-------------
Since the response status converter is the centralized class to handle
the decoding, it now WARN logs a message when it detects an unknown status
which will be converted into a generic FAILURE.

Result
------
Better debugging when an unknown status code is detected.

Change-Id: I8574eebd475a00b71c37a41876a18616860b9c97
Reviewed-on: http://review.couchbase.org/50156
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>