JVMCBC-192: Signal config reload on idle state key value event.
Motivation
----------
When no or not much traffic is going through the bucket, it can
happen that config changes are not properly picked up (especially
adding nodes, since on node removal the socket is closed which
the heuristics pick up).

Following the resiliency pattern of routine workouts, we need
to make sure that proper configs are available even if not
much load (or no load at all) is going through the system.

Modifications
-------------
IdleState events from netty are intercepted in the KeyValueHandler
and a config reload is signalled to the endpoint which in turn
will trigger it up to the configuration handler.

In the couchbase+cccp bucket case, a new configuration will be
proactively loaded from one of the nodes in the cluster.

Result
------
Cluster changes are picked up even when no load at all is going
through the system for the CCCP case. In all other cases those
events are streamed over HTTP which has the same effect.

Change-Id: I00c87049c80612c602b907d0818841151e5f0d76
Reviewed-on: http://review.couchbase.org/50177
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>