JVMCBC-195: Resubscribe HTTP Config Stream.
Motivation
----------
When a HTTP config stream is closed, the core needs to reconnect
if a subscription is still open to that bucket. This can happen
when the streaming node is failed over or removed and affects
memcached buckets on all server versions and couchbase buckets
on Server 2.2 and older.

Modifications
-------------
The code uses the observables repeatWhen clause to resubscribe
to the source observable if the stream is closed and a subscription
is still ongoing (the bucket has not been closed by the user).

Result
------
Better failure handling during failover or removal situations of
the attached http streaming node.

Change-Id: Ib83679d1e7d8421d33ee099a4976f5d5338ab54e
Reviewed-on: http://review.couchbase.org/50272
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>