JVMCBC-198: Populate BucketNodeLocator
Motivation
----------
Because the code uses the nodeLocator field as part of the config
to select the proper subclass in jackson, it is not accessible through
the regular json accessors.

Modifications
--------------
Since every bucket has a specific locator, it is okay to just pass the
actual locator type up to the abstract document.

An assertion has been added to the unit tests to make sure it is properly
populated.

Result
------
The configs now have their proper locator set and not null.

Change-Id: Idc42c3b478c823ac1ae50b09e8156b1abfcace82
Reviewed-on: http://review.couchbase.org/50716
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>