JVMCBC-203: replace loaded services Set with bitset in config.
Motivation
----------
As part of a larger effort to make the hot code path even faster,
we can make some hot methods even more performant.

Modifications
-------------
The Set which contains all the currently enabled service types
does not need to use HashSet lookups, instead it can be projected
onto a bitmap. The runtime lookup results in one logical AND which
can't be done any faster.

Result
------
Less code in the hot code path when looking up if the service
is enabled on the current configuration.

Change-Id: I5d2ade2c8416b4b696ec88d6bfd94b4dec4cd214
Reviewed-on: http://review.couchbase.org/51757
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>