JVMCBC-206: Allow to configure TCP_NODELAY
Motivation
----------
For better tuning of user's systems, it should be possible to
tune the TCP_NODELAY (which is enabled by default) as needed.

In addition, since this changeset is based on the new HELLO
command, it will ask the server to set the same, making sure
that whatever is chosen is set on both client and server.

Modifications
-------------
The option is now configurable on the environment as usual and
the changeset makes sure that the Endpoint is properly configured.

In addition the HELLO feature is set to the same value so if the
server supports it, it will be enabled/disabled.

Result
------
Tunable TCP_NODELAY on the client, and if supported, on the server
side.

Change-Id: Ie59b9c52dc7d1eb505de2e240fcff18695a38a6f
Reviewed-on: http://review.couchbase.org/51978
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>