JVMCBC-205: Clarify MutationToken
Motivation
----------
The MutationDescriptor, while accurate, should be replaced with a
more descriptive name: MutationToken. In addition, the vbucket id
needs to be added in order to be useful for RYOW.

Modifications
-------------
Renaming of the value object to MutationToken, as well as all related
methods and accessors.

The vbucket ID has been added as a member of the value object.

Result
------
Consistent naming across SDKs, as well as the addition of the vbucket
ID to the value object.

Change-Id: Ie1c23263cebf424051d855406a88246afefb45b8
Reviewed-on: http://review.couchbase.org/51999
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>