JVMCBC-201: Ignore IO pool errors on shutdown.
Motivation
----------
When onComplete is called on the subscriber, it can be the case that
it ends up on the computation pool, which is also shut down
concurrently.

Modifications
-------------
Errors are properly ignored during pool shutdown, since the only
important task is to get rid of the threads. Any race conditions
that throw errors and confuse users/spam logs are ignored on
purpose.

Result
------
No confusing errors on shutdown.

Change-Id: Ic3c210b2f89020128a9d6e9436f10f3027c69225
Reviewed-on: http://review.couchbase.org/52226
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>