JVMCBC-211: Reject replica gets to downed partition quickly.
Motivation
----------
When a replica get request comes along, and the client knows
based on its configuration that the vbucket is down (-1), it
can provide that information explicitly to the caller.

Modifications
-------------
Addition of a new ReplicaNotAvailableException, which is now
thrown instead of the ambiguous ReplicaNotConfiguredException
to clearly indicate that the replica partition is not
available at this point.

The locator has been adapted to throw the proper ones, depending
on the condition.

Result
------
The caller can quickly identify that the replica is not available,
for example to skip those and just return the replicas that
answered in the meantime.

Change-Id: I07cf29a662862620fda3558bf45ab15ed178895e
Reviewed-on: http://review.couchbase.org/52539
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>