JVMCBC-213: Publish a ConfigUpdatedEvent on the event bus
Motivation
----------
Certain applications might want to react every time a new configuration
is published, and it also helps during debugging and development sessions
(for example even if a rebalance is going on, no new config is published).

The event is not as spammy, since deduplication happens upfront, based
on the revid.

Modifications
-------------
A new event (ConfigUpdatedEvent) has been added and the RequestHandler
now publishes the event right after applying the configuration (which
takes precedence of course).

Result
------
New configuration events can be correlated and verified.

Change-Id: I6924ef882f275e9f58c092d87e8e782560455fbb
Reviewed-on: http://review.couchbase.org/52475
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>