JVMCBC-224: Do not let the CoreEnvironment.Builder implement the Env IF.
Motivation
----------
A constant source of confusion has been the fact that the CoreEnvironment
Builder also implements the CoreEnvironment. Also, this meant that methods
which are not supported on the builder needed to throw exception or (worse)
return null.

Modifications
-------------
This changeset removes the interface implementation and consequently gets
rid of the unneeded getters on the builder.

The environment code has been adapted to access the fields in the builder
directly, instead through its getter methods.

Result
------
The builder is now clearly distinguishable from the built environment
itself and also does not have to throw unsupported exceptions on methods
which are not meant to be called at all.

Change-Id: I0b1193c821025ea77012b3a9c9d6ce15a82d7a00
Reviewed-on: http://review.couchbase.org/53455
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>