JVMCBC-60: Add System Metric Collection and Events
Motivation
----------
It is important that vital JVM information like GC stats,
memory used/available and thread counts are gathered
and emitted as events from time to time so that continuous
profiling without extra tools is made easier.

This also lays the groundwork for logging events (and also
exposing them in many different ways) is possible at all.

Modifications
-------------
This changeset introduces the metrics namespace, which defines
a basic collector and its configuration. All this is designed
to be extensible and sub-classable if needed in the future.

Configurations can be overriden through the environment which
also does the state management (that is, unsubscribing the
interval subscriptions on shutdown).

The default time is 1 hour (to not spam the logs), but this is
completely tunable. An emit interval of 0 will disable the
gathering and emitting completely.

Result
------
System metric events are emitted.

Change-Id: I42ebd23358c406d899eeb991c105b82fb8779f4d
Reviewed-on: http://review.couchbase.org/53463
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>