Fix deadlocking environment test.
Motivation
----------
The trampoline scheduler is a poor choice if other code depends on
it since its executing on the same thread. Together with the recent
metric additions, a different scheduler is needed.

Modifications
-------------
The newThread scheduler does the job as well in this test case.

Result
------
No deadlock in the test.

Change-Id: I5d10b78441509456da461fe655932476d0f14aa5
Reviewed-on: http://review.couchbase.org/53541
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>