JVMCBC-228: Implement an Event Logging Consumer
Motivation
----------
For easy access, an event logger should be in place that can be used by the user
to log events which are incoming over the event bus.

Modifications
-------------
A LoggingConsumer has been implemented which consumes the event bus and logs
according to the specified log level and in the configured format.

To make things easier, a utility class has been added and all events now support
a versatile toMap() function that converts the objects into simple types. This
is then utilized to convert into a JSON representation if needed.

Also, if you just want to change the emit frequency some overloads have been added
to the configs to make it much simpler to work with.

Result
------
Easy logging of system events either as toString or through JSON.

Change-Id: I58f952010a9e75f0c3842f0a046ac73c095d1667
Reviewed-on: http://review.couchbase.org/53792
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>