JVMCBC-190: Add Support for Seqno-based Observe
Motivation
----------
In addition to CAS based observe, in combination with the
introduced mutation token it is now possible to perform much
more accurate observe logic, on the sequence number to also
detect concurrent mutations and handle failover scenarios
much more gracefully (and identify those scenarios).

Modifications
-------------
This changeset has three components to it, which slightly
overlap:

	1) The new observe for seqno message has been added and the
	   handlers and locators have been modified to respect it the
	   same as the cas-based observe command.
	2) Observe via cas has been refactored out in its own class
	   for easier change of the two implementations.
	3) observe via seqno has been added.

Note that by intention, the two different mechanisms share lots of code.
Since this is transparent, in the future this can be cleaned up where
most of the logic is moved to a parent class.

Result
------
Seqno based observe is possible if a mutation token is passed to the
Observe class. Otherwise it will transparently dispatch it to the
oldstyle cas observe.

Change-Id: I0866dad266c43ec28526d36e189f8fc7e7510f69
Reviewed-on: http://review.couchbase.org/54062
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>